Moving cmds to broker
